move the reflection tool to the utils/tools directory
